refactor: 패키지 구조 변경
- 변경된 REST API 요구사항에 따라 admin패키지를 제거
