+++ src/main/java/com/sjincho/delivery/food/dto/FoodUpdateRequest.java
@@ -1,14 +1,14 @@
-package com.sjincho.delivery.admin.food.dto;
+package com.sjincho.delivery.food.dto;
 
 import lombok.Getter;
 
 @Getter
-public class AdminFoodCreateRequest {
+public class FoodUpdateRequest {
     private final String name;
     private final String foodType;
     private final Long price;
 
-    public AdminFoodCreateRequest(final String name, final String foodType, final Long price) {
+    public FoodUpdateRequest(final String name, final String foodType, final Long price) {
         this.name = name;
         this.foodType = foodType;
         this.price = price;
