+++ src/main/java/com/sjincho/delivery/food/service/FoodService.java
@@ -1,5 +1,7 @@
 package com.sjincho.delivery.food.service;
 
+import com.sjincho.delivery.food.dto.FoodCreateRequest;
+import com.sjincho.delivery.food.dto.FoodUpdateRequest;
 import com.sjincho.delivery.exception.DeliveryApplicationException;
 import com.sjincho.delivery.exception.ErrorCode;
 import com.sjincho.delivery.food.domain.Food;
@@ -7,6 +9,8 @@ import com.sjincho.delivery.food.dto.FoodResponse;
 import com.sjincho.delivery.food.repository.FoodJpaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -32,5 +36,34 @@ public class FoodService {
         return foods.stream()
                 .map(FoodResponse::from)
                 .collect(Collectors.toList());
+    }
+
+    @Transactional
+    public Long register(final FoodCreateRequest request) {
+        final Food food = Food.create(request.getName(), request.getFoodType(), request.getPrice());
+
+        final Food saved = foodJpaRepository.save(food);
+
+        return saved.getId();
+    }
+
+    @Transactional
+    public FoodResponse update(final Long foodId, final FoodUpdateRequest request) {
+        final Food food = foodJpaRepository.findById(foodId).orElseThrow(() ->
+                new DeliveryApplicationException(ErrorCode.FOOD_NOT_FOUND, String.format("id:%d Not Found", foodId)));
+
+        food.update(request.getName(), request.getFoodType(), request.getPrice());
+
+        final Food updatedFood = foodJpaRepository.save(food);
+
+        return FoodResponse.from(updatedFood);
+    }
+
+    @Transactional
+    public void delete(final Long foodId) {
+        final Food food = foodJpaRepository.findById(foodId).orElseThrow(() ->
+                new DeliveryApplicationException(ErrorCode.FOOD_NOT_FOUND, String.format("id:%d Not Found", foodId)));
+
+        foodJpaRepository.delete(food);
     }
 }
